Work around namespace deficiency in PRANG 0.02
Changes which were introduced to PRANG after the version which got
converted from being a metaclass to a metaclass metarole (see
Moose::Cookbook, entries Meta::Recipe1 and Meta::Recipe2) are yet to be
brought forward, and will need to be for this to work.  This is required
to get the tests to run for the meantime.