Update to new PRANG role/plug-in API
It's no longer necessary/legal to specify xmlns => "*", which was a hack in
the first place.  It is now possible to have objects with the same nodeName
but different xmlns, so a lot of these extra properties are not necessary
now.  This also allows us to re-enable a test which had the same
root_element as another one.