Ignore another @lang attribute in the XML::Compare tests
These properties have a 'default' set in XML Schema, and the classes as
defined use the Moose 'default' property to make it always set; so, the
emitted version has the property set, and the read-in version doesn't.  As
with the epp:msg nodes.  Add the reason node to the list of ignored XPaths.